feat(desktop): strip archive UI from profile panel + add agent status badge
Phase 1 IXI-7 viewer ships clean -- destructive identity actions (archive,
persona-delete, identity surface) are deferred to a future holistic
agent-actions PR per tho. This commit:

- Removes Archive/Unarchive buttons, ProfileAdminSection (Admin/Danger
  zone), canArchive plumbing, archive flair display, and the
  useArchiveIdentityMutation/useUnarchiveIdentityMutation/
  useIsIdentityArchived/useOaOwnerQuery wiring from the profile panel.
- Drops desktop/tests/e2e/identity-archive.spec.ts (gate-matrix tests
  for the removed Archive button). identity-archive-hide.spec.ts stays
  -- it exercises the discovery-suppression contract (sidebar fold,
  mention filter, DM picker omission) which is independent UI.
- Extracts AgentStatusBadge from ManagedAgentRow.tsx into its own
  module so both the agents page and the profile panel render the
  same visual.
- Wires AgentStatusBadge into the profile Status field via a new
  optional displayNode on ProfileField, so the row shows a real pill
  (running / starting / stopped) instead of plain text. displayValue
  stays as the text fallback for tooltip + future copy support.
